requirements.txt
PySocks
pycryptodomex
requests
configparser
